# lib/apiservers/portlayer/swagger.json
{
	"swagger": "2.0",
	"basePath": "/",
	"info": {
		"description": "Port Layer API",
		"title": "Port Layer API",
		"version": "v0.0.1"
	},
	"produces": [
		"application/json"
	],
	"consumes": [
		"application/json"
	],
	"schemes": [
		"http"
	],
	"parameters": {
		"operationID": {
			"name": "Op-ID",
			"type": "string",
			"pattern": "[0-9]+\\.[0-9]+",
			"in": "header",
			"required": false
		}
	},
	"paths": {
...
		"/tasks": {
			"parameters": [
				{
					"$ref": "#/parameters/operationID"
				}
			],
			"get": {
				"description": "Initiates a task inspect operation",
				"summary": "Initiates a task inspect operation",
				"operationId": "Inspect",
				"tags": [
					"tasks"
				],
				"consumes": [
					"application/json",
					"application/octet-stream"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "config",
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/TaskInspectConfig"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/TaskInspectResponse"
						}
					},
					"404": {
						"description": "not found",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					},
            "409": {
              "description": "Task failed due to power state change",
              "schema": {
                  "$ref": "#/definitions/Error"
              }
            },
					"500": {
						"description": "Inspect of task failed",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					}
				}
			},
			"post": {
				"description": "Initiates a task join operation",
				"summary": "Initiates a task join operation",
				"operationId": "Join",
				"tags": [
					"tasks"
				],
				"consumes": [
					"application/json",
					"application/octet-stream"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "config",
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/TaskJoinConfig"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/TaskJoinResponse"
						}
					},
					"404": {
						"description": "not found",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					},
					"500": {
						"description": "Join of task failed",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					}
				}
			},
			"delete": {
				"description": "Initiates a task remove operation",
				"summary": "Initiates a task remove operation",
				"operationId": "Remove",
				"tags": [
					"tasks"
				],
				"consumes": [
					"application/json",
					"application/octet-stream"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "config",
						"in": "body",
						"schema": {
							"$ref": "#/definitions/TaskRemoveConfig"
						},
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/TaskRemoveResponse"
						}
					},
					"404": {
						"description": "not found",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					},
					"500": {
						"description": "Removal of task failed",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					}
				}
			}
		},
...
	},
	"definitions": {
		"Error": {
			"type": "object",
			"required": [
				"message"
			],
			"properties": {
				"code": {
					"type": "integer",
					"format": "int64"
				},
				"message": {
					"type": "string",
					"x-nullable": false
				}
			}
		},
...
		"ProcessConfig": {
			"type": "object",
			"properties": {
				"pid": {
					"type": "integer"
				},
				"execPath": {
					"type": "string"
				},
				"execArgs": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"workingDir": {
					"x-nullable": true,
					"type": "string"
				},
				"user": {
					"type": "string"
				},
				"env": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"status": {
					"type": "string"
				},
				"startTime": {
					"type": "integer",
					"format": "int64"
				},
				"stopTime": {
					"type": "integer",
					"format": "int64"
				},
				"exitCode": {
					"type": "integer",
					"format": "int32"
				},
				"errorMsg": {
					"type": "string"
				}
			}
		},
...
		"TaskJoinConfig": {
			"type": "object",
			"required": [
				"handle",
				"path"
			],
			"properties": {
				"handle": {
					"type": "object"
				},
				"id": {
					"type": "string",
					"x-nullable": false
				},
				"path": {
					"type": "string",
					"x-nullable": false
				},
				"args": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"workingDir": {
					"type": "string"
				},
				"env": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"user": {
					"type": "string"
				},
				"attach": {
					"type": "boolean",
					"default": true,
					"x-nullable": false
				},
				"openStdin": {
					"type": "boolean",
					"default": false,
					"x-nullable": false
				},
				"tty": {
					"type": "boolean",
					"default": false,
					"x-nullable": false
				},
				"stopSignal": {
					"type": "string"
				}
			}
		},
...
		"TaskRemoveConfig": {
			"type": "object",
			"required": [
				"handle",
				"id"
			],
			"properties": {
				"id": {
					"type": "string",
					"x-nullable": false
				},
				"handle": {
					"type": "object"
				}
			}
		},
		"TaskInspectConfig": {
			"type": "object",
			"required": [
				"handle",
				"id"
			],
			"properties": {
				"id": {
					"type": "string",
					"x-nullable": false
				},
				"handle": {
					"type": "object"
				}
			}
		},
...
		"TaskJoinResponse": {
			"type": "object",
			"required": [
				"handle",
				"id"
			],
			"properties": {
				"handle": {
					"type": "object"
				},
				"id": {
					"type": "string",
					"x-nullable": false
				}
			}
		},
...
		"TaskRemoveResponse": {
			"type": "object",
			"required": [
				"handle"
			],
			"properties": {
				"handle": {
					"type": "object"
				}
			}
		},
		"TaskInspectResponse": {
			"type": "object",
			"required": [
				"id"
			],
			"properties": {
				"id": {
					"type": "string",
					"x-nullable": false
				},
				"state": {
					"type": "string",
					"x-nullable": false
				},
				"exitCode": {
					"type": "integer",
					"x-nullable": false
				},
				"processConfig": {
					"$ref": "#/definitions/ProcessConfig"
				},
				"openStdin": {
					"type": "boolean",
					"default": false,
					"x-nullable": false
				},
				"openStderr": {
					"type": "boolean",
					"default": false,
					"x-nullable": false
				},
				"openStdout": {
					"type": "boolean",
					"default": false,
					"x-nullable": false
				},
				"tty": {
					"type": "boolean",
					"default": false,
					"x-nullable": false
				},
				"user": {
					"type": "string"
				},
				"pid": {
					"type": "integer",
					"x-nullable": false
				}
			}
		}
	}
}